moved RegisterPage.tsx to RegisterPage folder
