+++ src/pages/RegisterPage/RegisterPage.tsx
@@ -1,11 +1,11 @@
-import '../pages/LoginPage/_login_page.scss';
-import { Logo } from '../components/Logo/Logo';
+import '../LoginPage/_login_page.scss';
+import { Logo } from '../../components/Logo/Logo';
 import { useSelector } from "react-redux";
-import { selectCurrentUser } from "../store/auth/authSlice";
+import { selectCurrentUser } from "../../store/auth/authSlice";
 import {useEffect} from "react";
-import {navigateToDefaultRoute} from "../utils/navigation/navigation";
+import {navigateToDefaultRoute} from "../../utils/navigation/navigation";
 import {useNavigate} from "react-router";
-import {RegisterForm} from "../components/Formik/Forms/RegisterForm";
+import {RegisterForm} from "../../components/Formik/Forms/RegisterForm";
 
 
 export const RegisterPage = () => {
